modal for delete operation

+++ components/screens/UploadFile/FilesTable.tsx
@@ -4,6 +4,10 @@ import { FileUpload } from "@prisma/client";
 import { useUser } from "@/utils/hooks/useUser";
 import Image from "next/image";
 import { useAppSelector } from "@/utils/redux/store";
+import { Spinner } from "@/components/elements/Loaders";
+import TableLoader from "@/components/elements/TableLoader";
+import ConfirmationModal from "@/components/elements/DeleteModal"; // Import the ConfirmationModal component
+import { Button } from "@/components/elements/Button";
 
 interface TableProps extends React.HTMLProps<HTMLTableCellElement> {
   children: React.ReactNode;
@@ -43,27 +47,38 @@ const FilesTable = () => {
   const [files, setFiles] = useState<FileUpload[]>([]);
   const { user } = useUser();
   const searchTerm = useAppSelector((state) => state.searchTerm.term);
-
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [showConfirmationModal, setShowConfirmationModal] = useState(false);
+  const [fileToDelete, setFileToDelete] = useState<FileUpload | null>(null);
   if (!user) {
     throw new Error("User not found");
   }
 
-  const handleDelete = async (id: string, userId: string, fileName: string) => {
-    if (user) {
+  const handleDelete = async (file: FileUpload) => {
+    setFileToDelete(file);
+    setShowConfirmationModal(true);
+  };
+
+  const confirmDelete = async () => {
+    if (user && fileToDelete) {
       try {
         const response = await fetch("/api/upload/file-upload", {
           method: "DELETE",
           headers: {
             "Content-Type": "application/json",
           },
-          body: JSON.stringify({ id, userId, fileName }),
+          body: JSON.stringify({
+            id: fileToDelete.id,
+            userId: fileToDelete.userId,
+            fileName: fileToDelete.fileName,
+          }),
         });
         console.log(response);
         if (response.ok) {
           console.log("File deleted successfully!");
-          // setFiles((prevFiles) =>
-          //   prevFiles.filter((file) => file.id !== userId)
-          // );
+          setFiles((prevFiles) =>
+            prevFiles.filter((file) => file.id !== fileToDelete.id)
+          );
         } else {
           console.error("Failed to delete file:", response.statusText);
         }
@@ -71,12 +86,20 @@ const FilesTable = () => {
         console.error("Error deleting file:", error);
       }
     }
+    setShowConfirmationModal(false);
+    setFileToDelete(null);
+  };
+
+  const cancelDelete = () => {
+    setShowConfirmationModal(false);
+    setFileToDelete(null);
   };
 
   useEffect(() => {
     if (user) {
       const fetchFiles = async () => {
         try {
+          setIsLoading(true);
           const response = await fetch(
             `/api/upload/file-upload?userId=${user.id}`
           );
@@ -87,6 +110,8 @@ const FilesTable = () => {
           setFiles(filesData);
         } catch (error) {
           console.error("Error fetching files:", error);
+        } finally {
+          setIsLoading(false); // Set loading state to false after fetching data
         }
       };
 
@@ -100,8 +125,30 @@ const FilesTable = () => {
       )
     : files;
 
+  if (filteredFiles.length === 0 && searchTerm) {
+    return (
+      <div className="text-center mt-4 font-semibold text-lg">
+        No document found
+      </div>
+    );
+  }
+
+  if (isLoading) {
+    return (
+      <div className="px-6">
+        <TableLoader />;
+      </div>
+    );
+  }
   return (
     <div className="overflow-x-auto mx-3 mt-6">
+      <ConfirmationModal
+        isOpen={showConfirmationModal}
+        onConfirm={confirmDelete}
+        onCancel={cancelDelete}
+        title="Confirm Delete"
+        message="Are you sure you want to delete this file?"
+      />
       <table className="rounded-xl table-auto min-w-full divide-y divide-gray-200">
         <thead>
           <tr className="bg-gray-100">
@@ -138,7 +185,7 @@ const FilesTable = () => {
                   : file.fileUrl}
               </TableCell>
               <TableCell>
-                <span className="flex justify-center items-center w-1/3 bg-green-200 text-green-600 px-2 py-1 rounded-full text-xs">
+                <span className="flex justify-center items-center w-3/4 bg-green-200 text-green-600 px-2 py-1 rounded-full text-xs">
                   {file.mimeType &&
                     file.mimeType.replace(
                       /(?:application|image|video|text)\/(.+)/,
@@ -151,15 +198,26 @@ const FilesTable = () => {
                   new Date(file.createdAt).toLocaleDateString("en-GB")}
               </TableCell>
               <TableCell className="cursor-pointer">
-                <Image
-                  src={`/icons/delete.svg`}
-                  width={20}
-                  height={20}
-                  alt="Delete Icon"
-                  onClick={() =>
-                    handleDelete(file.id!, file.userId!, file.fileName!)
-                  }
-                />
+                <Button
+                  className="border px-2 py-1 border-gray-400 bg-transparent text-black rounded-md hover:bg-gray-300"
+                  alt="Delete Button"
+                  onClick={() => handleDelete(file)}
+                >
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth={1.5}
+                    stroke="currentColor"
+                    className="w-6 h-6"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"
+                    />
+                  </svg>
+                </Button>
               </TableCell>
             </tr>
           ))}
